Path based deployment of GeoServer in Tomcat
This ensures that the GeoServer application is hosted under a given
path, instead of always being hosted under /geoserver.
The path is defined in the environment variable APP_PATH_PREFIX in
the form foo#bar#, which leads the application being
hosted under /foo/bar/geoserver/. If the env var is not set the
GeoServer will be hosted under /geoserver as usual.